src/presentation/controller/signup-controller.ts
import { AddAccount, Authentication } from '@domain/usecases'
import { EmailInUseError } from '../errors'
import { badRequest, forbidden, serverError, successResponse } from '../helpers'
import { Controller, HttpResponse, Validation } from '../protocols'

export class SignUpController implements Controller {
  constructor(
    private readonly addAccount: AddAccount,
    private readonly validation: Validation,
    private readonly authentication: Authentication
  ) {}

  public async handle(
    request: SignUpController.Request
  ): Promise<HttpResponse> {
    try {
      const error = this.validation.validate(request)

      if (error) {
        return badRequest(error)
      }

      const { password, email, name } = request

      const isValid = await this.addAccount.add({ email, name, password })

      if (!isValid) {
        return forbidden(new EmailInUseError())
      }
      const authenticationModel = await this.authentication.auth({
        email,
        password
      })

      return successResponse(authenticationModel)
    } catch (error) {
      return serverError(error)
    }
  }
}

export namespace SignUpController {
  export type Request = {
    name: string
    email: string
    password: string
    passwordConfirmation: string
  }
}
